replace provider clients components with vNext implementation
