+++ bitwarden_license/bit-web/src/app/admin-console/providers/clients/clients.component.html
@@ -1,5 +1,9 @@
 <app-header>
-  <bit-search [placeholder]="'search' | i18n" [(ngModel)]="searchText"></bit-search>
+  <bit-search
+    class="tw-grow"
+    [formControl]="searchControl"
+    [placeholder]="'search' | i18n"
+  ></bit-search>
   <a bitButton routerLink="create" *ngIf="manageOrganizations" buttonType="primary">
     <i class="bwi bwi-plus bwi-fw" aria-hidden="true"></i>
     {{ "newClient" | i18n }}
@@ -24,63 +28,55 @@
   <span class="tw-sr-only">{{ "loading" | i18n }}</span>
 </ng-container>
 
-<ng-container
-  *ngIf="!loading && (clients | search: searchText : 'organizationName' : 'id') as searchedClients"
->
-  <p *ngIf="!searchedClients.length">{{ "noClientsInList" | i18n }}</p>
-  <ng-container *ngIf="searchedClients.length">
-    <table
-      class="table table-hover table-list"
-      infiniteScroll
-      [infiniteScrollDistance]="1"
-      [infiniteScrollDisabled]="!isPaging()"
-      (scrolled)="loadMore()"
+<ng-container *ngIf="!loading">
+  <p *ngIf="dataSource.data.length < 1">{{ "noClientsInList" | i18n }}</p>
+  <ng-container *ngIf="dataSource.data.length >= 1">
+    <bit-table-scroll
+      [dataSource]="dataSource"
+      [rowSize]="53"
+      class="tw-table tw-w-full table-hover table-list"
     >
-      <thead>
-        <tr>
-          <th colspan="2">{{ "name" | i18n }}</th>
-          <th>{{ "numberOfUsers" | i18n }}</th>
-          <th>{{ "billingPlan" | i18n }}</th>
-          <th></th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr *ngFor="let o of searchedClients">
-          <td width="30">
-            <bit-avatar [text]="o.organizationName" [id]="o.id" size="small"></bit-avatar>
-          </td>
-          <td>
-            <a [routerLink]="['/organizations', o.organizationId]">{{ o.organizationName }}</a>
-          </td>
-          <td>
-            <span>{{ o.userCount }}</span>
-            <span *ngIf="o.seats != null"> / {{ o.seats }}</span>
-          </td>
-          <td>
-            <span>{{ o.plan }}</span>
-          </td>
-          <td class="table-list-options" *ngIf="manageOrganizations">
-            <div class="dropdown" appListDropdown>
-              <button
-                class="btn btn-outline-secondary dropdown-toggle"
-                type="button"
-                data-toggle="dropdown"
-                aria-haspopup="true"
-                aria-expanded="false"
-                appA11yTitle="{{ 'options' | i18n }}"
-              >
-                <i class="bwi bwi-cog bwi-lg" aria-hidden="true"></i>
-              </button>
-              <div class="dropdown-menu dropdown-menu-right">
-                <a class="dropdown-item text-danger" href="#" appStopClick (click)="remove(o)">
-                  <i class="bwi bwi-fw bwi-close" aria-hidden="true"></i>
-                  {{ "remove" | i18n }}
-                </a>
-              </div>
+      <ng-container header>
+        <th bitCell colspan="2" bitSortable="organizationName">{{ "name" | i18n }}</th>
+        <th bitCell bitSortable="seats">{{ "numberOfUsers" | i18n }}</th>
+        <th bitCell bitSortable="plan">{{ "billingPlan" | i18n }}</th>
+        <th bitCell></th>
+      </ng-container>
+      <ng-template bitRowDef let-row>
+        <td bitCell width="30">
+          <bit-avatar [text]="row.organizationName" [id]="row.id" size="small"></bit-avatar>
+        </td>
+        <td bitCell width="325">
+          <a [routerLink]="['/organizations', row.organizationId]">{{ row.organizationName }}</a>
+        </td>
+        <td bitCell>
+          <span>{{ row.userCount }}</span>
+          <span *ngIf="row.seats != null"> / {{ row.seats }}</span>
+        </td>
+        <td bitCell width="150">
+          <span>{{ row.plan }}</span>
+        </td>
+        <td class="table-list-options" *ngIf="manageOrganizations">
+          <div class="dropdown" appListDropdown>
+            <button
+              class="btn btn-outline-secondary dropdown-toggle"
+              type="button"
+              data-toggle="dropdown"
+              aria-haspopup="true"
+              aria-expanded="false"
+              appA11yTitle="{{ 'options' | i18n }}"
+            >
+              <i class="bwi bwi-cog bwi-lg" aria-hidden="true"></i>
+            </button>
+            <div class="dropdown-menu dropdown-menu-right">
+              <a class="dropdown-item text-danger" href="#" appStopClick (click)="remove(row)">
+                <i class="bwi bwi-fw bwi-close" aria-hidden="true"></i>
+                {{ "remove" | i18n }}
+              </a>
             </div>
-          </td>
-        </tr>
-      </tbody>
-    </table>
+          </div>
+        </td>
+      </ng-template>
+    </bit-table-scroll>
   </ng-container>
 </ng-container>
