+++ bitwarden_license/bit-web/src/app/admin-console/providers/providers.module.ts
@@ -11,9 +11,7 @@ import { OssModule } from "@bitwarden/web-vault/app/oss.module";
 
 import {
   CreateClientDialogComponent,
-  NoClientsComponent,
   ManageClientNameDialogComponent,
-  ManageClientsComponent,
   ManageClientSubscriptionDialogComponent,
   ProviderBillingHistoryComponent,
   ProviderSubscriptionComponent,
@@ -21,7 +19,6 @@ import {
 } from "../../billing/providers";
 
 import { AddOrganizationComponent } from "./clients/add-organization.component";
-import { ClientsComponent } from "./clients/clients.component";
 import { CreateOrganizationComponent } from "./clients/create-organization.component";
 import { AcceptProviderComponent } from "./manage/accept-provider.component";
 import { AddEditMemberDialogComponent } from "./manage/dialogs/add-edit-member-dialog.component";
@@ -59,7 +56,6 @@ import { VerifyRecoverDeleteProviderComponent } from "./verify-recover-delete-pr
     AddOrganizationComponent,
     BulkConfirmDialogComponent,
     BulkRemoveDialogComponent,
-    ClientsComponent,
     CreateOrganizationComponent,
     EventsComponent,
     MembersComponent,
@@ -68,8 +64,6 @@ import { VerifyRecoverDeleteProviderComponent } from "./verify-recover-delete-pr
     UserAddEditComponent,
     AddEditMemberDialogComponent,
     CreateClientDialogComponent,
-    NoClientsComponent,
-    ManageClientsComponent,
     ManageClientNameDialogComponent,
     ManageClientSubscriptionDialogComponent,
     ProviderBillingHistoryComponent,
