+++ bitwarden_license/bit-web/src/app/billing/providers/clients/index.ts
@@ -3,5 +3,4 @@ export * from "./manage-clients.component";
 export * from "./manage-client-name-dialog.component";
 export * from "./manage-client-subscription-dialog.component";
 export * from "./no-clients.component";
-export * from "./vnext-manage-clients.component";
 export * from "./replace.pipe";
