+++ bitwarden_license/bit-web/src/app/billing/providers/clients/manage-clients.component.ts
@@ -1,23 +1,28 @@
-// FIXME: Update this file to be type safe and remove this and next line
-// @ts-strict-ignore
 import { Component } from "@angular/core";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
+import { FormControl } from "@angular/forms";
 import { ActivatedRoute, Router } from "@angular/router";
 import { firstValueFrom, from, lastValueFrom, map } from "rxjs";
-import { switchMap } from "rxjs/operators";
+import { debounceTime, first, switchMap } from "rxjs/operators";
 
-import { SearchService } from "@bitwarden/common/abstractions/search.service";
 import { ProviderService } from "@bitwarden/common/admin-console/abstractions/provider.service";
 import { ProviderStatusType, ProviderUserType } from "@bitwarden/common/admin-console/enums";
 import { Provider } from "@bitwarden/common/admin-console/models/domain/provider";
 import { ProviderOrganizationOrganizationDetailsResponse } from "@bitwarden/common/admin-console/models/response/provider/provider-organization.response";
-import { BillingApiServiceAbstraction as BillingApiService } from "@bitwarden/common/billing/abstractions/billing-api.service.abstraction";
+import { BillingApiServiceAbstraction } from "@bitwarden/common/billing/abstractions";
 import { PlanResponse } from "@bitwarden/common/billing/models/response/plan.response";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { ValidationService } from "@bitwarden/common/platform/abstractions/validation.service";
-import { DialogService, ToastService } from "@bitwarden/components";
+import {
+  AvatarModule,
+  DialogService,
+  TableDataSource,
+  TableModule,
+  ToastService,
+} from "@bitwarden/components";
+import { SharedOrganizationModule } from "@bitwarden/web-vault/app/admin-console/organizations/shared";
+import { HeaderModule } from "@bitwarden/web-vault/app/layouts/header/header.module";
 
-import { BaseClientsComponent } from "../../../admin-console/providers/clients/base-clients.component";
 import { WebProviderService } from "../../../admin-console/providers/services/web-provider.service";
 
 import {
@@ -32,47 +37,53 @@ import {
   ManageClientSubscriptionDialogResultType,
   openManageClientSubscriptionDialog,
 } from "./manage-client-subscription-dialog.component";
+import { NoClientsComponent } from "./no-clients.component";
+import { ReplacePipe } from "./replace.pipe";
 
 @Component({
   templateUrl: "manage-clients.component.html",
+  standalone: true,
+  imports: [
+    AvatarModule,
+    TableModule,
+    HeaderModule,
+    SharedOrganizationModule,
+    NoClientsComponent,
+    ReplacePipe,
+  ],
 })
-export class ManageClientsComponent extends BaseClientsComponent {
-  providerId: string;
-  provider: Provider;
-
+export class ManageClientsComponent {
+  providerId: string = "";
+  provider: Provider | undefined;
   loading = true;
   isProviderAdmin = false;
+  dataSource: TableDataSource<ProviderOrganizationOrganizationDetailsResponse> =
+    new TableDataSource();
 
-  protected plans: PlanResponse[];
+  protected searchControl = new FormControl("", { nonNullable: true });
+  protected plans: PlanResponse[] = [];
 
   constructor(
-    private billingApiService: BillingApiService,
+    private billingApiService: BillingApiServiceAbstraction,
     private providerService: ProviderService,
     private router: Router,
-    activatedRoute: ActivatedRoute,
-    dialogService: DialogService,
-    i18nService: I18nService,
-    searchService: SearchService,
-    toastService: ToastService,
-    validationService: ValidationService,
-    webProviderService: WebProviderService,
+    private activatedRoute: ActivatedRoute,
+    private dialogService: DialogService,
+    private i18nService: I18nService,
+    private toastService: ToastService,
+    private validationService: ValidationService,
+    private webProviderService: WebProviderService,
   ) {
-    super(
-      activatedRoute,
-      dialogService,
-      i18nService,
-      searchService,
-      toastService,
-      validationService,
-      webProviderService,
-    );
+    this.activatedRoute.queryParams.pipe(first(), takeUntilDestroyed()).subscribe((queryParams) => {
+      this.searchControl.setValue(queryParams.search);
+    });
 
-    this.activatedRoute.parent.params
-      .pipe(
+    this.activatedRoute.parent?.params
+      ?.pipe(
         switchMap((params) => {
           this.providerId = params.providerId;
           return this.providerService.get$(this.providerId).pipe(
-            map((provider) => provider?.providerStatus === ProviderStatusType.Billable),
+            map((provider: Provider) => provider?.providerStatus === ProviderStatusType.Billable),
             map((isBillable) => {
               if (!isBillable) {
                 return from(
@@ -89,20 +100,24 @@ export class ManageClientsComponent extends BaseClientsComponent {
         takeUntilDestroyed(),
       )
       .subscribe();
-  }
 
-  removeMonthly = (plan: string) => plan.replace(" (Monthly)", "");
+    this.searchControl.valueChanges
+      .pipe(debounceTime(200), takeUntilDestroyed())
+      .subscribe((searchText) => {
+        this.dataSource.filter = (data) =>
+          data.organizationName.toLowerCase().indexOf(searchText.toLowerCase()) > -1;
+      });
+  }
 
   async load() {
     this.provider = await firstValueFrom(this.providerService.get$(this.providerId));
 
-    this.isProviderAdmin = this.provider.type === ProviderUserType.ProviderAdmin;
+    this.isProviderAdmin = this.provider?.type === ProviderUserType.ProviderAdmin;
 
-    this.clients = (
-      await this.billingApiService.getProviderClientOrganizations(this.providerId)
-    ).data;
+    const clients = (await this.billingApiService.getProviderClientOrganizations(this.providerId))
+      .data;
 
-    this.dataSource.data = this.clients;
+    this.dataSource.data = clients;
 
     this.plans = (await this.billingApiService.getPlans()).data;
 
@@ -131,7 +146,7 @@ export class ManageClientsComponent extends BaseClientsComponent {
         organization: {
           id: organization.id,
           name: organization.organizationName,
-          seats: organization.seats,
+          seats: organization.seats ? organization.seats : 0,
         },
       },
     });
@@ -149,7 +164,7 @@ export class ManageClientsComponent extends BaseClientsComponent {
     const dialogRef = openManageClientSubscriptionDialog(this.dialogService, {
       data: {
         organization,
-        provider: this.provider,
+        provider: this.provider!,
       },
     });
 
@@ -159,4 +174,28 @@ export class ManageClientsComponent extends BaseClientsComponent {
       await this.load();
     }
   };
+
+  async remove(organization: ProviderOrganizationOrganizationDetailsResponse) {
+    const confirmed = await this.dialogService.openSimpleDialog({
+      title: organization.organizationName,
+      content: { key: "detachOrganizationConfirmation" },
+      type: "warning",
+    });
+
+    if (!confirmed) {
+      return;
+    }
+
+    try {
+      await this.webProviderService.detachOrganization(this.providerId, organization.id);
+      this.toastService.showToast({
+        variant: "success",
+        title: "",
+        message: this.i18nService.t("detachedOrganization", organization.organizationName),
+      });
+      await this.load();
+    } catch (e) {
+      this.validationService.showError(e);
+    }
+  }
 }
