+++ bitwarden_license/bit-web/src/app/billing/providers/clients/no-clients.component.ts
@@ -1,6 +1,7 @@
 import { Component, EventEmitter, Input, Output } from "@angular/core";
 
 import { svgIcon } from "@bitwarden/components";
+import { SharedOrganizationModule } from "@bitwarden/web-vault/app/admin-console/organizations/shared";
 
 const gearIcon = svgIcon`
 <svg width="120" height="120" viewBox="0 0 120 120" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -23,6 +24,8 @@ const gearIcon = svgIcon`
 
 @Component({
   selector: "app-no-clients",
+  standalone: true,
+  imports: [SharedOrganizationModule],
   template: `<div class="tw-flex tw-flex-col tw-items-center tw-text-info">
     <bit-icon [icon]="icon"></bit-icon>
     <p class="tw-mt-4">{{ "noClients" | i18n }}</p>
